requirements.txt
django==4.2
pillow
python-dotenv
google-generativeai
pdf2image
streamlit